Client.getRoomList has been broken for some time.
I looked over the tfm source code (THANK YOU TOCU) and there is an extra useless Packet.readBoolean().
I've looked over the packet buffers generated by the client and it seems like yes, there is an extra byte at the end that is not read. Hopefully you can also look over it, but this completely fixes it for me!